requirements.txt
psycopg2-binary==2.8.2
prettytable==0.7.2